requirements.txt
coverage
djangocms-admin-style
django-appconf
djangocms-helper
